Fixed some problems assoc. with texture source and dst rects, added some new reduced textures, fixed a variety of compiler warnings throughout.

+++ SpaceGamePrelim/Scene/Scene.cpp
@@ -14,7 +14,7 @@ Scene::~Scene()
 void Scene::Update()
 {
 
-	for (int i = 0; i < m_Objects.size(); i++)
+	for (size_t i = 0; i < m_Objects.size(); i++)
 	{
 		m_Objects[i]->Update();
 	}
@@ -22,7 +22,7 @@ void Scene::Update()
 }
 void Scene::Render()
 {
-	for (int i = 0; i < m_Objects.size(); i++)
+	for (size_t i = 0; i < m_Objects.size(); i++)
 	{
 		m_Objects[i]->Draw();
 	}
